SP-359 refactor: service 디렉터리 구조 변
