--- src/main/java/com/ludo/study/studymatchingplatform/study/service/ParticipantService.java
@@ -1,25 +1,0 @@
-package com.ludo.study.studymatchingplatform.study.service;
-
-import org.springframework.stereotype.Service;
-
-import com.ludo.study.studymatchingplatform.study.domain.Participant;
-import com.ludo.study.studymatchingplatform.study.domain.Position;
-import com.ludo.study.studymatchingplatform.study.domain.Role;
-import com.ludo.study.studymatchingplatform.study.domain.Study;
-import com.ludo.study.studymatchingplatform.study.repository.ParticipantRepositoryImpl;
-import com.ludo.study.studymatchingplatform.user.domain.User;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class ParticipantService {
-
-	private final ParticipantRepositoryImpl participantRepository;
-
-	public void add(final Study study, final User user, final Position position, final Role role) {
-		final Participant participant = Participant.from(study, user, position, role);
-		study.addParticipant(participantRepository.save(participant));
-	}
-
-}
