+++ src/main/java/com/ludo/study/studymatchingplatform/study/service/study/participant/ParticipantService.java
@@ -1,0 +1,25 @@
+package com.ludo.study.studymatchingplatform.study.service.study.participant;
+
+import org.springframework.stereotype.Service;
+
+import com.ludo.study.studymatchingplatform.study.domain.recruitment.position.Position;
+import com.ludo.study.studymatchingplatform.study.domain.study.Study;
+import com.ludo.study.studymatchingplatform.study.domain.study.participant.Participant;
+import com.ludo.study.studymatchingplatform.study.domain.study.participant.Role;
+import com.ludo.study.studymatchingplatform.study.repository.study.participant.ParticipantRepositoryImpl;
+import com.ludo.study.studymatchingplatform.user.domain.user.User;
+
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+public class ParticipantService {
+
+	private final ParticipantRepositoryImpl participantRepository;
+
+	public void add(final Study study, final User user, final Position position, final Role role) {
+		final Participant participant = Participant.from(study, user, position, role);
+		study.addParticipant(participantRepository.save(participant));
+	}
+
+}
